Reviwed #166
changed the label: does not seem that ''agroforestry fallow'' term exist. There is only mentions of improved fallow in litterature.
Also, I delete the link the shifting cultivation. In shifting cultivation, the fallow phase is not managed at all (and might not include any trees) whereas in the improved fallow it is.